+++ src/main/kotlin/com/ravit/stash/controller/externals/ai/response/AiChatExternalResponse.kt
@@ -2,4 +2,5 @@ package com.ravit.stash.controller.externals.ai.response
 
 data class AiChatExternalResponse(
     val response: String,
+    val isError: Boolean = false,
 )